refactor : setup controller
