+++ lib/cube/cube_state.dart
@@ -1,5 +1,6 @@
 import 'package:cy_cube/cube/cube_controller/arrange_controller.dart';
 import 'package:cy_cube/cube/cube_controller/extension_controller.dart';
+import 'package:cy_cube/cube/cube_controller/setup_controller.dart';
 import 'package:cy_cube/cube/cube_view/cube_component.dart';
 import 'cube_model/single_cube_model.dart';
 import 'package:flutter/material.dart';
@@ -236,138 +237,17 @@ class CubeState {
     cubeModels[edgeIds[3]].component = edgeComponent;
   }
 
-  void _setupSingleFace({
-    required List<int> cubeIDs,
-    required List<SingleCubeComponentFaceModel> cubeFaces,
-    required Facing facing,
-  }) {
-    int i = 0;
-    for (int id in cubeIDs) {
-      Map<Facing, Color> updatedCubeColor =
-          Map.from(cubeModels[id].component.cubeColor);
-      updatedCubeColor[facing] = cubeFaces[i].color!;
-      cubeModels[id] = SingleCubeModel(
-        component: CubeComponent(
-          cubeColor: updatedCubeColor,
-        ),
-        x: cubeModels[id].x,
-        y: cubeModels[id].y,
-        z: cubeModels[id].z,
-      );
-      i++;
-    }
-  }
-
   void setupCubeWithScanningColor(
       List<List<SingleCubeComponentFaceModel>> cubeFaces) {
-    for (int i = 0; i < cubeFaces.length; i++) {
-      if (cubeFaces[i][4].color == Colors.white) {
-        _setupSingleFace(
-          cubeIDs: [24, 25, 26, 15, 16, 17, 6, 7, 8],
-          cubeFaces: cubeFaces[i],
-          facing: Facing.top,
-        );
-      } else if (cubeFaces[i][4].color == Colors.yellow) {
-        _setupSingleFace(
-          cubeIDs: [0, 1, 2, 9, 10, 11, 18, 19, 20],
-          cubeFaces: cubeFaces[i],
-          facing: Facing.down,
-        );
-      } else if (cubeFaces[i][4].color == Colors.red) {
-        _setupSingleFace(
-          cubeIDs: [20, 11, 2, 23, 14, 5, 26, 17, 8],
-          cubeFaces: cubeFaces[i],
-          facing: Facing.right,
-        );
-      } else if (cubeFaces[i][4].color == Colors.orange) {
-        _setupSingleFace(
-          cubeIDs: [0, 9, 18, 3, 12, 21, 6, 15, 24],
-          cubeFaces: cubeFaces[i],
-          facing: Facing.left,
-        );
-      } else if (cubeFaces[i][4].color == Colors.blue) {
-        _setupSingleFace(
-          cubeIDs: [2, 1, 0, 5, 4, 3, 8, 7, 6],
-          cubeFaces: cubeFaces[i],
-          facing: Facing.back,
-        );
-      } else {
-        _setupSingleFace(
-          cubeIDs: [18, 19, 20, 21, 22, 23, 24, 25, 26],
-          cubeFaces: cubeFaces[i],
-          facing: Facing.front,
-        );
-      }
-    }
+    SetupController.setupCubeWithScanningColor(cubeFaces);
   }
 
-  static String transformColorToString(Color color) {
-    if (color == Colors.red) {
-      return 'red';
-    } else if (color == Colors.orange) {
-      return 'orange';
-    } else if (color == Colors.white) {
-      return 'white';
-    } else if (color == Colors.yellow) {
-      return 'yellow';
-    } else if (color == Colors.blue) {
-      return 'blue';
-    } else if (color == Colors.black) {
-      return 'black';
-    } else {
-      return 'green';
-    }
+  List<String> generateCubeStatus() {
+    return SetupController.generateCubeStatus();
   }
 
-  static Color _transformStringToColor(String color) {
-    if (color == 'red') {
-      return Colors.red;
-    } else if (color == 'orange') {
-      return Colors.orange;
-    } else if (color == 'white') {
-      return Colors.white;
-    } else if (color == 'yellow') {
-      return Colors.yellow;
-    } else if (color == 'blue') {
-      return Colors.blue;
-    } else if (color == 'black') {
-      return Colors.black;
-    } else {
-      return Colors.green;
-    }
-  }
-
-  List<String> outputCubeState() {
-    List<String> cubeStatus = [];
-    for (Facing cubeFace in cubeFaces) {
-      for (int cubeFaceID in cubeFaceIDs[cubeFace]!) {
-        String cubeColor = transformColorToString(
-            cubeModels[cubeFaceID].component.cubeColor[cubeFace]!);
-        cubeStatus.add(cubeColor);
-      }
-    }
-    return cubeStatus;
-  }
-
-  void setCubeState({required List<String> cubeStatus}) {
-    int index = 0;
-    for (Facing cubeFace in cubeFaces) {
-      List<SingleCubeComponentFaceModel> cubeFaceModes = [];
-      for (int cubeFaceID in cubeFaceIDs[cubeFace]!) {
-        SingleCubeComponentFaceModel cubeFaceModel =
-            SingleCubeComponentFaceModel(
-          id: cubeFaceID,
-          color: _transformStringToColor(cubeStatus[index++]),
-          isSelected: false,
-        );
-        cubeFaceModes.add(cubeFaceModel);
-      }
-      _setupSingleFace(
-        cubeIDs: cubeFaceIDs[cubeFace]!,
-        cubeFaces: cubeFaceModes,
-        facing: cubeFace,
-      );
-    }
+  void setCubeStatus({required List<String> cubeStatus}) {
+    SetupController.setCubeStatus(cubeStatus: cubeStatus);
   }
 
   SingleCubeFace show2DFace({required Facing facing}) {
